Add Parser::expect_any
A common pattern is to peek for a token kind and then get the token.
This became a hassle where every post-peek pull on the lexer had to
repeat the same set of unwrapping.  Put all this unwrapping in a
function that panics (which should never happen because it's been
checked already).